==> notebooks/0-xt-parse_datasets.ipynb <==
import os
import pandas as pd
from itertools import islice

### 1 - Load and parse helper files for loading raw OPD data file

data_path = os.path.join('..', 'data', 'raw')

#helpers = []
#for file in os.listdir(data_path):
#    if file.endswith('.sps'):
#        helpers.append(pd.read_csv(os.path.join(data_path, file)))

with open(os.path.join(data_path, 'opd2011.sps'), 'r', encoding='windows-1252') as f:
    # Keep the trunk after 'DATA LIST...' and before the first '.'
    for i, line in enumerate(f):
        if line.startswith('DATA LIST'):
            start = i + 1
        if line.startswith('.'):
            end = i
            break

    #helper = [line.strip() for line in helper[start:end]]
    f.seek(0)
    helper = list(islice(f, start, end))
    helper = [line.strip() for line in helper]

# Convert the helper to a DataFrame, split each line by whitespace
helper = pd.DataFrame([line.replace(' (A)', '').split(maxsplit=1) for line in helper]).rename(columns={0:'name'})

# Remove the rows that ends with digits followed by 'D'
helper = helper[~helper['name'].str.contains(r'\dD$')]

helper['width'] = helper[1].apply(lambda x: abs(eval(x)) + 1 if ('-' in x) else 1)

opd = pd.read_fwf(os.path.join(data_path, 'opd2011'), widths=helper['width'], names=helper['name'], header=None)
opd

opd.iloc[:, 20:40]

opd.PROC1R.value_counts()
