==> notebooks/2.2-jy-xt-nlp.ipynb <==
import pandas as pd
import os
import json

### 1 - Load the training data

file_path = os.path.join('..', 'data', 'cleaned')

train_df = pd.read_csv(os.path.join(file_path, 'train.csv'), low_memory=False)
train_df.head()

# Load the variables dictionary
with open(os.path.join(file_path, 'variables.json'), 'r') as f:
    variables = json.load(f)

variables

train_df[variables['physicianDiagnoses']]
