==> notebooks/4.0-jy-xt-user_input_matching.ipynb <==
import pandas as pd
import numpy as np
import os
import json

## 1 - Load data

### 1.1 - Load the preprocessed data

processed_file_path = os.path.join('..', 'data', 'processed')
clustering_model_name = 'KMeans'
processed_file_name = f'train_{clustering_model_name}.csv'

train_df = pd.read_csv(os.path.join(processed_file_path, processed_file_name), low_memory=False)
train_df.head()

### 1.2 - Load the vairiables dictionary

# Load the variables dictionary
variables_file_path = os.path.join('..', 'data', 'cleaned')

with open(os.path.join(variables_file_path, 'variables.json'), 'r') as f:
    variables = json.load(f)

print(f'Variable Categories:\n')
for category, list in variables.items():
    print(f'{category}')
    print(f'{list}')

## 2 - Preprocess user input

# Load custom function to combine text features
import sys
sys.path.append('../src/features/')

from combine_textual import combine_features


# Define the list of features to combine
feature_list = [
    'AGE', 'SEX', 'USETOBAC', 
    'MAJOR', 'RFV1', 'RFV2', 'RFV3', 
    'BMI', 'TEMPF', 'BPSYS', 'BPDIAS',
    'ARTHRTIS', 'ASTHMA', 'CANCER', 'CEBVD', 'CHF', 'CRF', 'COPD', 'DEPRN', 'DIABETES', 'HYPLIPID', 'HTN', 'IHD', 'OBESITY', 'OSTPRSIS', 'NOCHRON', 'DMP',
    'DIAG1', 'DIAG2', 'DIAG3'
]

user_input = combine_features(train_df.iloc[0].copy(), feature_list[:-3])

## 3 - Generate vectors and compute similarities/relevances

# Load model directly
from transformers import AutoTokenizer, AutoModel
import torch

if torch.cuda.is_available():
    device = 'cuda'
elif torch.backends.mps.is_available():
    device= 'mps'
else:
    device = 'cpu'
print(f'Using device: {device}')

model_name = "microsoft/BiomedNLP-BiomedBERT-base-uncased-abstract-fulltext"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModel.from_pretrained(model_name).to(device)

def get_embedding(text):
    encoded_input = tokenizer(text, return_tensors='pt', padding=True, truncation=True, max_length=512).to(device)

    with torch.no_grad():
        output = model(**encoded_input)
        sentence_embedding = output.last_hidden_state[:, 0, :]
        #sentence_embedding = output.pooler_output
    return sentence_embedding


def compare_to_groups(user_input, records_by_group, batch_size=32):
    # preprocess user input
    #user_input = preprocess_text(user_input)

    # Print the user input
    print(f'User Input:')
    print(f'{user_input}')
    print()

    user_embedding = get_embedding(user_input)

    group_relevances = {}
    for group_id, records in records_by_group.items():
        batch_embeddings = []
        for i in range(0, len(records), batch_size):
            batch_records = records[i:i+batch_size]
            batch_embeddings.append(get_embedding([text for text in batch_records]))

        group_embeddings = torch.cat(batch_embeddings, dim=0)

        group_similarities = torch.nn.functional.cosine_similarity(user_embedding, group_embeddings, dim=1)

        # Print the max and min similarity scores and the corresponding texts within the group
        print(f'Group {group_id}:')
        print(f'Max Similarity: {group_similarities.max().item()}')

        max_index = group_similarities.argmax().item()
        print(f'Most Similar Record:')
        print(f'{records[max_index]}')
        print()

        print(f'Min Similarity: {group_similarities.min().item()}')

        min_index = group_similarities.argmin().item()
        print(f'Least Similar Record:')
        print(f'{records[min_index]}')
        print()
        print()
    
        group_relevance = group_similarities.mean(dim=0).item()
        group_relevances[group_id] = group_relevance

    return group_relevances

train_df['ProcessedText'] = train_df.apply(combine_features, args=(feature_list[:-3],), axis=1)

records_by_group = train_df.groupby('cluster')['ProcessedText'].apply(lambda x: x.tolist()).to_dict()
records_by_group.keys()

compare_to_groups(user_input, records_by_group)
